% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_credentials.R
\name{gesis_credentials}
\alias{gesis_credentials}
\title{Saves GESIS credentials}
\usage{
gesis_credentials(user, password)
}
\arguments{
\item{user}{your user name.}

\item{password}{your password.}
}
\description{
\code{save_credentials} saves a set of GESIS credentials using the \code{keyring} package.
}
\details{
User/password are stored in Keychain on macOS, Credential Store on Windows or Secret Service API on Linux. 
 If a user/password pair for a database already exists, it is silently replaced with the new pair.  
 This function relies on the \code{\link{keyring}}  package.
}
\examples{

 \dontrun{
 gesis_credentials(
		user="stein.rokkan@sowi.uni-mannheim.de", 
		password="comparativepolitics")
 }


}
\seealso{
\code{\link{keyring}}
}
