% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gesis_download_data.R
\name{gesis_download_data}
\alias{gesis_download_data}
\title{Get a data set from GESIS}
\usage{
gesis_download_data(za, purpose, file = NULL, type = "dta", print = TRUE)
}
\arguments{
\item{za}{the study number (starts with "ZA")}

\item{purpose}{select a purpose for the download (see details)}

\item{file}{file (and path) to use for storage (detault: [za].[type])}

\item{type}{file type to download (default: STATA dta)}

\item{print}{print summary of server response (default) or return response object?}
}
\value{
only returns httr response object if \code{print=FALSE}.
}
\description{
\code{gesis_download_data} downloads data from the GESIS data repository.
}
\details{
You need to define a purpose for downloading a data set. The options accepted by the 
the GESIS API are as follows:

\itemize{
 \item scientific_research: for scientific research (incl. doctorate)
 \item commercial_research: for research with commercial mandate
 \item lecturer: in a course as lecturer
 \item final_thesis: for final thesis of the study programme (e.g. Bachelor/Master thesis)
 \item further_education: for further education and qualification
 \item non_scientific: for non-scientific purposes
 \item studies: in the course of my studies
}

File options vary across studies. Typically, studies are available as 
por, dta, sps, sav file.
}
\examples{

\dontrun{
Download ZA0078: Attitudes towards Europe, 1962
gesis_download_data(za="ZA0078", purpose="scientific_research")
}



}
